protect routes

+++ routes/web.php
@@ -17,16 +17,16 @@ use Illuminate\Support\Facades\Route;
 */
 
 Route::get('/', [ListingController::class, 'index']);
-Route::get('/listings/create', [ListingController::class, 'create']);
-Route::post('/listings', [ListingController::class, 'store']);
-Route::get('/listings/{listing}/edit', [ListingController::class, 'edit']);
-Route::delete('/listings/{listing}', [ListingController::class, 'delete']);
+Route::get('/listings/create', [ListingController::class, 'create'])->middleware('auth');
+Route::post('/listings', [ListingController::class, 'store'])->middleware('auth');
+Route::get('/listings/{listing}/edit', [ListingController::class, 'edit'])->middleware('auth');
+Route::delete('/listings/{listing}', [ListingController::class, 'delete'])->middleware('auth');
 Route::get('/listings/{listing}', [ListingController::class, 'show']);
 
-Route::get('/register', [UserController::class, 'create']);
+Route::get('/register', [UserController::class, 'create'])->middleware('guess');
 Route::post('/users', [UserController::class, 'store']);
-Route::post('/logout', [UserController::class, 'logout']);
-Route::get('/login', [UserController::class, 'login']);
+Route::post('/logout', [UserController::class, 'logout'])->middleware('auth');
+Route::get('/login', [UserController::class, 'login'])->name('login')->middleware('guess');
 Route::post('/users/authenticate', [UserController::class, 'authenticate']);
